Add database updating and removing functionality
I basically added 2 commands. The first one checks which servers are up or down and stores the indexes of the latter in an array. The second one removes all servers in that array if provided with no parameters (or you can provide a string as a parameter and the server with that name will be removed)